// packages/server/src/enums.ts
export enum Equipments {
  BANDS = 'Bands',
  BARBELL = 'Barbell',
  BODY_READ_ONLYMACHINE = 'Body read onlymachine',
  CABLE = 'Cable',
  DUMBBELL = 'Dumbbell',
  E_Z_CURL_BAR = 'E z curl bar',
  EXERCISE_BALL = 'Exercise ball',
  FOAM_ROLL = 'Foam roll',
  KETTLEBELLS = 'Kettlebells',
  MEDICINE_BALL = 'Medicine ball',
  NONE = 'None',
  OTHER = 'Other',
}

export enum FitnessLevel {
  Advanced = 'Advanced',
  Beginner = 'Beginner',
  Intermediate = 'Intermediate',
}

export enum Forms {
  Powder = 'Powder',
  Capsule = 'Capsule',
  Tablet = 'Tablet',
  Liquid = 'Liquid',
  Bars = 'Bars',
  Other = 'Other',
}

export enum Roles {
  admin = 'admin',
  editor = 'editor',
  user = 'user',
}

export enum AdsLocations {
  SLIDER = 'Slider',
  MEGA_MENU = 'Mega Menu',
  HOME_SECTION = 'Home Section',
  POPUP = 'Popup',
}

export enum Muscles {
  FOREARM_BRACHIORADIALIS = 'Forearm brachioradialis',
  GLUTES_GLUTEUS_MAXIMUS_AND_MEDIUS = 'Glutes gluteus maximus and medius',
  HAMSTRINGS_BICEPS_FEMORIS = 'Hamstrings biceps femoris',
  TRICEPS_TRICEPS_BRACHII = 'Triceps triceps brachii',
  ABS_RECTUS_ABDOMINIS = 'Abs rectus abdominis',
  BICEPS_BICEPS_BRACHII = 'Biceps biceps brachii',
  CALVES_GASTROCNEMIUS = 'Calves gastrocnemius',
  CHEST_PECTORALIS = 'Chest pectoralis',
  LATS_LATISSIMUS_DORSI = 'Lats latissimus dorsi',
  LOWER_BACK = 'Lower back',
  MIDDLE_BACK_RHOMBOIDS = 'Middle back rhomboids',
  NECK = 'Neck',
  QUADS_QUADRICEPS = 'Quads quadriceps',
  SHOULDERS_DELTOIDS = 'Shoulders deltoids',
  TRAPS_TRAPEZIUS = 'Traps trapezius',
}

export enum CategoryTypes {
  articles = 'articles',
  exercises = 'exercises',
  store = 'store',
}

export enum LanguageCode {
  ar = 'ar',
  en = 'en',
}

export enum Title {
  Athlete = 'Athlete',
  Doctor = 'Doctor',
  InternationalNutritionist = 'International Nutritionist',
  Nutritionist = 'Nutritionist',
  Pt = 'Pt',
  Trainer = 'Trainer',
}

export enum CartStatuses {
  Waiting = 'Waiting',
  Processing = 'Processing',
  Delivered = 'Delivered',
  InProgress = 'InProgress',
}

export enum Gender {
  Male = 'Male',
  Female = 'Female',
}

export enum Goals {
  Bulk = 'Bulk',
  Endurance = 'Endurance',
  FatLoss = 'Fat loss',
  Fitness = 'Fitness',
  MassGain = 'Mass gain',
  MuscleMass = 'Muscle mass',
  MuscleStrength = 'Muscle strength',
  Shred = 'Shred',
}